Remove the limit on number of individual function arguments, replace `u8` with `u32`.

+++ src/compiler/parser/optimizer.rs
@@ -197,7 +197,7 @@ impl Expr {
     fn is_partial(self: &Self, nargs: usize) -> bool {
         match &self.1 {
             ExprType::NativeFunction(native) => match native.nargs() {
-                Some(expected_nargs) => expected_nargs > nargs as u8 && nargs > 0,
+                Some(expected_nargs) => expected_nargs > nargs as u32 && nargs > 0,
                 None => false,
             },
             _ => false
